refactor: STEP2 피드백 사항 반영 #5

+++ OpenMarket/OpenMarket/Model/NetWork/ParsingType/ProductDetailAsk.swift
@@ -7,9 +7,9 @@ enum ProductDetailAsk {
         let name: String
         let thumbnail: String
         let currency: Currency
-        let price: Int
-        let bargainPrice: Int
-        let discountedPrice: Int
+        let price: Double
+        let bargainPrice: Double
+        let discountedPrice: Double
         let stock: Int
         let images: [Image]
         let vendors: Vendor
