+++ OpenMarket/OpenMarket/Model/NetWork/ParsingType/ProductModify.swift
@@ -5,9 +5,9 @@ enum ProductModify {
         let name: String?
         let description: String?
         let thumbnailId: Int?
-        let price: Int?
+        let price: Double?
         let currency: Currency?
-        let discountedPrice: Int?
+        let discountedPrice: Double?
         let stock: Int?
         let secret: String
         
@@ -24,9 +24,9 @@ enum ProductModify {
         let name: String
         let thumbnail: String
         let currency: Currency
-        let price: Int
-        let bargainPrice: Int
-        let discountedPrice: Int
+        let price: Double
+        let bargainPrice: Double
+        let discountedPrice: Double
         let stock: Int
         let images: [Image]
         let vendors: Vendor
